chore: delete dead population-signal path
The `signal_source: "population"` alternative architecture (commit 41969af,
"Horizon separation") was built alongside research-based signals but never
switched on in production. Research path won. 1088 LOC of parallel code.

Deleted: signal_generator.py, population_reader.py, technical_scorer.py,
test_signal_generator.py. Moved read_predictions into signal_reader.py
(still needed for GBM prediction capture in rationale). Collapsed
signal_source branching in main.py to research-only. Dropped stale
`trading:` block from both example and private risk.yaml.

Part of the 2026-04-13 system audit (alpha-engine-audit-system-260413.md),
PR A in the dead-code deletion batch.
